# apps/server/internal/docs/swagger.yaml
basePath: /
definitions:
...
  online.batchLogoutRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: Internal admin platform API documentation.
  title: Admin Service API
  version: 0.1.0
paths:
...
  /v1/monitor/logs/operations:
    get:
      description: 按标题、业务类型、状态等分页查询
      parameters:
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 模块标题
        in: query
        name: title
        type: string
      - description: 业务类型
        in: query
        name: businessType
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 操作人员
        in: query
        name: operName
        type: string
      - description: 请求方法
        in: query
        name: requestMethod
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取操作日志
      tags:
      - Monitor/OperLog
  /v1/monitor/logs/operations/{id}:
    delete:
      description: 根据ID删除操作日志
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除操作日志
      tags:
      - Monitor/OperLog
    get:
      description: 根据ID查看操作日志
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取操作日志详情
      tags:
      - Monitor/OperLog
  /v1/monitor/online/users:
    get:
      description: 按用户名、IP、上线时间过滤在线会话
      parameters:
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 用户名
        in: query
        name: userName
        type: string
      - description: IP地址
        in: query
        name: ipaddr
        type: string
      - description: 开始时间(RFC3339)
        in: query
        name: since
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取在线用户
      tags:
      - Monitor/Online
  /v1/monitor/online/users/{id}:
    get:
      description: 根据会话ID查询在线用户
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取在线用户详情
      tags:
      - Monitor/Online
  /v1/monitor/online/users/{id}/force-logout:
    post:
      description: 根据会话ID强制注销在线用户
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 强制下线单个会话
      tags:
      - Monitor/Online
  /v1/monitor/online/users/batch-logout:
    post:
      consumes:
      - application/json
      description: 批量注销多个在线用户
      parameters:
      - description: 会话ID集合
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/online.batchLogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 批量下线会话
      tags:
      - Monitor/Online
  /v1/monitor/server:
    get:
      description: 获取 CPU、内存、磁盘等服务器指标
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 服务器监控
      tags:
      - Monitor/Server
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"